Cleanup: extract RecordingData and RecordingStatus from TrackRecordingService.
